[fs] nfs: have flexfiles mirror keep creds for both ro and rw layouts
Message-id: <2794dd4ffe07513c1646063a9ca670dc92d3bcb8.1469703539.git.bcodding@redhat.com>
Patchwork-id: 156047
O-Subject: [RHEL7 PATCH fs 12/12] nfs: have flexfiles mirror keep creds for both ro and rw layouts
Bugzilla: 1358386
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@poochiereds.net>

A mirror can be shared between multiple layouts, even with different
iomodes. That makes stats gathering simpler, but it causes a problem
when we get different creds in READ vs. RW layouts.

The current code drops the newer credentials onto the floor when this
occurs. That's problematic when you fetch a READ layout first, and then
a RW. If the READ layout doesn't have the correct creds to do a write,
then writes will fail.

We could just overwrite the READ credentials with the RW ones, but that
would break the ability for the server to fence the layout for reads if
things go awry. We need to be able to revert to the earlier READ creds
if the RW layout is returned afterward.

The simplest fix is to just keep two sets of creds per mirror. One for
READ layouts and one for RW, and then use the appropriate set depending
on the iomode of the layout segment.

Also fix up some RCU nits that sparse found.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Anna Schumaker <Anna.Schumaker@Netapp.com>
Upstream-id: 3064b6861d00b7124558cdf79f9387f948361be3
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1358386
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11402387
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>