[scsi] cxlflash: Verify problem state area is mapped before notifying shutdown
Message-id: <20160729173537.41608.57447.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 156133
O-Subject: [PATCH RHEL7.3 BZ1361359 6/6] cxlflash: Verify problem state area is mapped before notifying shutdown
Bugzilla: 1361359
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1361359

Description:
------------
If an EEH or some other hard error occurs while the adapter instance was
being initialized, on the subsequent shutdown of the device, the system
could crash with:

[c000000f1da03b60] c0000000005eccfc pci_device_shutdown+0x6c/0x100
[c000000f1da03ba0] c0000000006d67d4 device_shutdown+0x1b4/0x2c0
[c000000f1da03c40] c0000000000ea30c kernel_restart_prepare+0x5c/0x80
[c000000f1da03c70] c0000000000ea48c kernel_restart+0x2c/0xc0
[c000000f1da03ce0] c0000000000ea970 SyS_reboot+0x1c0/0x2d0
[c000000f1da03e30] c000000000009204 system_call+0x38/0xb4

This crash is due to the AFU not being mapped when the shutdown
notification routine is called and is a regression that was inserted
recently with Commit 704c4b0ddc03 ("cxlflash: Shutdown notify support
for CXL Flash cards").

As a fix, shutdown notification should only occur when the AFU is
mapped.

Fixes: 704c4b0ddc03 ("cxlflash: Shutdown notify support for CXL Flash cards")
Signed-off-by: Uma Krishnan <ukrishn@linux.vnet.ibm.com>
Reviewed-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Acked-by: Matthew R. Ochs <mrochs@linux.vnet.ibm.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>

Upstream Status:
----------------
https://git.kernel.org/cgit/linux/kernel/git/next/linux-next.git/commit/?id=1bd2b2823e5aa219b2442fd4568a075efe81de5f

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>