[x86] perf: Fix copy_from_user_nmi() return if range is not ok
Message-id: <1470226960-5477-4-git-send-email-jolsa@redhat.com>
Patchwork-id: 156392
O-Subject: [PATCH RHEL7.3 BZ1361670 3/3] perf/x86: Fix copy_from_user_nmi() return if range is not ok
Bugzilla: 1361670
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1361670
https://bugzilla.redhat.com/show_bug.cgi?id=1361670

upstream
========
commit ebf2d2689de551d90965090bb991fc640a0c0d41
Author: Yann Droneaud <ydroneaud@opteya.com>
Date:   Mon Jun 22 21:38:43 2015 +0200

description
===========
Commit 0a196848ca36 ("perf: Fix arch_perf_out_copy_user default"),
changes copy_from_user_nmi() to return the number of
remaining bytes so that it behave like copy_from_user().

Unfortunately, when the range is outside of the process
memory, the return value  is still the number of byte
copied, eg. 0, instead of the remaining bytes.

As all users of copy_from_user_nmi() were modified as
part of commit 0a196848ca36, the function should be
fixed to return the total number of bytes if range is
not correct.

Signed-off-by: Rafael Aquini <aquini@redhat.com>