[x86] smpboot: Fix cpu_init_udelay=10000 corner case boot parameter misbehavior
Message-id: <20160816160050.20041.55204.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 156966
O-Subject: [PATCH RHEL7.3 BZ1365413 5/7] x86/smpboot: Fix cpu_init_udelay=10000 corner case boot parameter misbehavior
Bugzilla: 1365413
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1365413

Description:
------------
For legacy machines cpu_init_udelay defaults to 10,000.
For modern machines it is set to 0.

The user should be able to set cpu_init_udelay to
any value on the cmdline, including 10,000.

Before this patch, that was seen as "unchanged from default"
and thus on a modern machine, the user request was ignored
and the delay was set to 0.

Signed-off-by: Len Brown <len.brown@intel.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: dparsons@brightdsl.net
Cc: shrybman@teksavvy.com
Link: http://lkml.kernel.org/r/de363cdbbcfcca1d22569683f7eb9873e0177251.1444968087.git.len.brown@intel.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=f1ccd249319efca4ee4faf1d904f5a362cac7c81

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>