[x86] tsc: Enumerate SKL cpu_khz and tsc_khz via CPUID
Message-id: <1471520593-1021-1-git-send-email-prarit@redhat.com>
Patchwork-id: 157076
O-Subject: [RHEL7.3 PATCH BZ 1366396] x86/tsc: Enumerate SKL cpu_khz and tsc_khz via CPUID
Bugzilla: 1366396
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1366396

Build info:
Date: Wed Aug 17 21:36:55 EDT 2016
Build OS: Red Hat Enterprise Linux Server release 7.3 Beta (Maipo)
System name: intel-purley-01.ml3.eng.bos.redhat.com with -j56
Built on: kernel-3.10.0-493.el7
Arch built: x86_64

This patch is required for RHEL7.3 Purley 2S support.

This took a while to debug and figure out -- the bottom line is that it is
again possible for the TSC and CPU to have different values on modern CPUs.

I suspect this is due to a side effect of HWP & Turbo mode, however, I have not
been able to prove that conclusively.  I'm also surprised we haven't hit this
issue before.  The 2S (2 socket) Purley system fails to do an initial
calibration of the TSC with

	tsc: Unable to calibrate against PIT

While not fatal to the TSC, it is a message that is surprising to see on a
modern system.  These patches take into account the possible difference and
the PIT again will calibrate the TSC correctly.  I've tested this on both
Purleys, a Skylake, and a few other random systems.

commit aa297292d708e89773b3b2cdcaf33f01bfa095d8
Author: Len Brown <len.brown@intel.com>
Date:   Fri Jun 17 01:22:51 2016 -0400

    x86/tsc: Enumerate SKL cpu_khz and tsc_khz via CPUID

    Skylake CPU base-frequency and TSC frequency may differ
    by up to 2%.

    Enumerate CPU and TSC frequencies separately, allowing
    cpu_khz and tsc_khz to differ.

    The existing CPU frequency calibration mechanism is unchanged.
    However, CPUID extensions are preferred, when available.

    CPUID.0x16 is preferred over MSR and timer calibration
    for CPU frequency discovery.

    CPUID.0x15 takes precedence over CPU-frequency
    for TSC frequency discovery.

    Signed-off-by: Len Brown <len.brown@intel.com>
    Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
    Cc: Linus Torvalds <torvalds@linux-foundation.org>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Link: http://lkml.kernel.org/r/b27ec289fd005833b27d694d9c2dbb716c5cdff7.1466138954.git.len.brown@intel.com
    Signed-off-by: Ingo Molnar <mingo@kernel.org>

Cc: Jiri Olsa <jolsa@redhat.com>
Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Vitaly Kuznetsov <vkuznets@redhat.com>
Cc: sbest@redhat.com
Cc: darcari@redhat.com
Signed-off-by: Rafael Aquini <aquini@redhat.com>