[fs] pmem: disable dax mounting in the prsence of media errors
Message-id: <20160818212216.19618-6-jmoyer@redhat.com>
Patchwork-id: 157168
O-Subject: [RHEL7 PATCH 5/5] pmem: disable dax mounting in the prsence of media errors
Bugzilla: 1367132
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

BZ: 1367132

This patch is RHEL-only.  Upstream can handle dax mounts with media
errors, but RHEL lacks the MCE-safe memcpy infrastructure.  In
order to prevent infinite reboots when there is a media error, and
an application is unlucky enough to always touch the errant region,
we simply disable mounting a file system with "-o dax" when there
are known errors.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>