[include] bluetooth: Fix kabi breakage in struct hci_core
Message-id: <1472232718-59458-1-git-send-email-dzickus@redhat.com>
Patchwork-id: 157349
O-Subject: [RHEL7 PATCH] bluetooth: Fix kabi breakage in struct hci_core
Bugzilla: 1370583
RH-Acked-by: Sabrina Dubroca <sdubroca@redhat.com>
RH-Acked-by: Hannes Frederic Sowa <hannes@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1370583
Brew: http://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11667247

My bluetooth rebase to 4.5 accidentally broke the kabi in a subtle way.  It was
not detected by our normal tools because the way GENKSYMS hides the real code.

Instead Jerome's new kabi tool detected the size increase and reported it to
me.  Thanks Jerome!

What happened was an two embedded structs (discovery and conn_hash) grew during
the rebase.  To workaround the kabi breakage due to the size increase, I
deprecated the old values and added the new larger structs to the end.  The
goal was to leave the old sized structs in the original spot to maintain the
hole needed for kabi complaince (but just rename the elements to avoid usage).

However, I didn't do this right.  Instead the deprecated elements still utilize
the new size and not the old because I did not snapshot the original struct and
its size.

Fixing this isn't as easy as I would have hoped.  Kabi macros were floated
around to assist with this problem.  But there was an added twist: embedded
enums.

As a result the complexities of the macros to fix this made the whole fix
difficult to understand, so I kept it simple for now.

To understand this fix, you have to follow this logic:

 -There is an _orig struct and a non _orig struct (the new expanded one).
 -The new expanded struct has GENKSYMS wrapped around the expanded elements.
- When checking kabi, GNEKSYMS is defined and we use the new expanded structs
  but with the expanded elements undefined to retain kabi compatibility.
- When compiling, the old elements use the _orig structs to retain correct size
  and the new elements use the new expanded struct with the expanded elements
  defined.

Functionally nothing should change other than a smaller hci_conn_hash struct.

Signed-off-by: Rafael Aquini <aquini@redhat.com>