[fs] gfs2: Fix extended attribute readahead optimization
Message-id: <756516896.11479554.1472568702708.JavaMail.zimbra@redhat.com>
Patchwork-id: 157451
O-Subject: [RHEL7.3 PATCH] [Addendum] gfs2: Fix extended attribute readahead optimization
Bugzilla: 1256539
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Andreas Gruenbacher <agruenba@redhat.com>

Hi,

This patch is an addendum to the two patches already appearing in the
RHEL7.3 kernel for bz 1256539. A regression was found during testing,
and this patch was written to fix the regression. Andreas Gruenbacher
wrote the original patch; I ported it to the RHEL7 kernel.

It was built into a RHEL7.3 kernel here:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11677810

It was tested on the performance team's "afc" cluster using SAS Calibration,
which is where we caught the regression.

This is a RHEL7 port of this upstream patch:
https://git.kernel.org/cgit/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2/meta_io.c?h=for-next&id=23e5671a79be00b2f1c895aa93ff40fb75a4647e

rhbz#1256539

Patch description:

Commit 39b0555f didn't check for a failing bio_add_page in
gfs2_submit_bhs. This could cause I/O requests to get lost, and the
affected buffer heads to stay locked forever.  Fix that by submitting
the current bio and allocating another one when bio_add_page fails.  (It
is guaranteed that we can at least add one page to a bio.)

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Bob Peterson <rpeterso@redhat.com>
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>