[x86] mm: non-linear virtual memory fix for KNL4 erratum
Message-id: <1473220476-13988-2-git-send-email-lwoodman@redhat.com>
Patchwork-id: 157667
O-Subject: [REHL7.3 PATCH 1/2] Fix conflicts between Knights Landing A/D fix and SOFT_DIRTY support.
Bugzilla: 1372047
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Mainline does not have real support for remap_file_pages()
(nonlinear memory) any more.  We now just fake it with linear
VMAs on top of the old API.  But, RHEL7 has the original code,
which our "KNL4 erratum" fix failed to patch.

RHEL7 marks non-present nonlinear PTEs with a special bit to
differentiate them from regular PTEs.  Unforunately, that bit
happens to be _PAGE_DIRTY, which can get set as a result of the
"KNL4 erratum".

A nonlinear PTE for the first page in a file has this value:
0x0000000000000040, which happens to be one of the values that
can be erroneously set on cleared PTEs with the erratum.

We need to find another bit.

On 32-bit non-PAE, we have no free bits, and the code is in sorry
shape anyway.  We do not have the erratum there, so just do not
move the bit around there.  (Not sure if this even matters for
RHEL7).

Fortunatley, we have lots of bits on 64-bit.  Just move it one
up from bit 6 to bit 7 (_PAGE_BIT_PSE).

I think the "#if _PAGE_BIT_FILE < _PAGE_BIT_PROTNONE" check was
dead code.  But, preserve it and add an error just in case.  Also,
open-code the SWP_TYPE_BITS just like we do in mainline.  The
Derivation from _PAGE_BIT_FILE and _PAGE_BIT_PRESENT was probably
another bug, and was confusing anyway.

So the PTE bits now look like this:

     bitnrs: |     ...            | 11| 10|  9|8|7|6|5| 4| 3|2|1|0|
     names:  |     ...            |SW3|SW2|SW1|G|L|D|A|CD|WT|U|W|P|
     swppte: | OFFSET (14-63)  |  TYPE (9-13) |0|X|X|X| X| X|X|X|0|
     ptenon: | PHYS ADDR (12-63)  |  X|  X|  X|1|X|?|X| X| X|X|X|0|
     nonlin: |    OFFSET (12-63)  |  0|  0|  0|0|0|1|0| 0| 0|0|0|0| <- before
     nonlin: |    OFFSET (12-63)  |  0|  0|  0|0|1|0|0| 0| 0|0|0|0| <- after

I am not 100% if we ever have PROT_NONE nonlinear ptes.  But, if we
do, this arrangement will support it.

Signed-off-by: Rafael Aquini <aquini@redhat.com>