[x86] mm: Fix regression panic at boot time seen on some NUMA systems
Message-id: <1473220476-13988-3-git-send-email-lwoodman@redhat.com>
Patchwork-id: 157666
O-Subject: [REHL7.3 PATCH 2/2] Fix regression panic at boot time seen on some NUMA systems.
Bugzilla: 1372047
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Due to a conflict with SOFT_DIRTY support in 7.3, the fix for
Intel Knights Landing A/D leak (BZ1347159) requires the additional
patch attached here.  This was discovered when trying to determine
why a test case failed on a 7.2.z test kernel (BZ1365475) and not 7.3.
Here are the details, courtesy of Dave Hansen.

The soft dirty code sets one more bit in the PTE (_PAGE_SOFT_DIRTY)
in additionto any other bits being set.  So, a nonlinear pte has
_PAGE_FILE and_PAGE_SOFT_DIRTY with the SOFT_DIRTY config option
enabled.  Without it(or in a kernel without soft dirty support like
RHEL7.2), it just has _PAGE_FILE set.

_PAGE_FILE is aliased to _PAGE_DIRTY, one of the bits affected by
the KNL erratum and, as such, is masked off when checking a pte for
being pte_none().  But _PAGE_SOFT_DIRTY is not affected by the
erratum and not masked off.

The RHEL7.2 bug here (with the erratum in place) comes in when the
__mm_populate() code called by remap_file_pages() goes and attempts
to fault in all the pte_none() ptes.  A PTE with just _PAGE_DIRTY
(aka _PAGE_FILE) is assumed to be caused by the erratum and pte_none()
(incorrectly) returns true and the PTE is overwritten (repopulated)
with a linearly-generated PTE.  This linear PTE where we want a
nonlinear one is what LKP notices.

But, the soft-dirty code *also* sets _PAGE_SOFT_DIRTY.  That extra bit
(not being affected by the erratum and not being masked off) causes
pte_none() to (correctly) return false, and the code functions as
expected, accidentally.

So, the SOFT_DIRTY support, in the common case at least, adds an extra
bit to the remap_file_pages() PTEs and works around the underlying bug.
But, there's still an underlying bug and probably some other ways to hit
it on RHEL7.3.  SOFT_DIRTY just happened to cover up the evidence in
the once place we had a nice test case to trigger the bug.

Signed-off-by: Rafael Aquini <aquini@redhat.com>