[iommu] vt-d: Disable passthrough mode on Kexec kernel
Message-id: <20160908161917.23859.20330.stgit@amt.stowe>
Patchwork-id: 157728
O-Subject: [RHEL7.3 PATCH] iommu/vt-d: Disable passthrough mode on Kexec kernel
Bugzilla: 1367621
RH-Acked-by: Xunlei Pang <xpang@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1367621
Upstream Status: RHEL only!!!

This patch is an attempt to avoid the need for an upstream series since
it is too large to consider this late in the release cycle and would also
be too large to consider for a subsequent z-stream fix.  The upstream
series that addresses this issue was introduced by Joerg Roedel during
v4.3:
  https://lkml.org/lkml/2015/8/5/516
  [PATCH 00/26] iommu/vt-d: Clean up device<->domain attachment
  Specifically [PATCH 08/26] - iommu/vt-d: Don't pre-allocate domain ids
                               for si_domain

The upstream series only incidentally avoids this issue by removing the
requirement that the si_domain has the same domain id across all IOMMUs
basically, this area was re-designed by the series mentioned and thus
inadvertently fixed the error).

While this patch/fix will be RHEL specific (which I generally frown on
very much) it is constrained, only coming into play when all three
conditions - platform with two or more IOMMUs, "iommu=pt" is specified on
the command line, and a crash kernel is Kexec'ed - are met.

We are not aware of a requirement to support passthrough mode on the Kexec
kernel since passthrough is primarily used to maintain host I/O device
performance while still enabling device assignment.  The Kexec kernel
copies the data structures from the previous IOMMU configuration so
devices will remain in a si_domain context until the Kexec kernel driver
initializes the device and performs DMA at which point the device is moved
to a domain of its own.

Signed-off-by: Myron Stowe <myron.stowe@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>