[powerpc] kvm: book3s: Don't crash if irqfd used with no in-kernel XICS emulation
Message-id: <1473833966-25009-2-git-send-email-dgibson@redhat.com>
Patchwork-id: 157811
O-Subject: [RHEL-7.3 kernel PATCH 1/2] KVM: PPC: Book3S: Don't crash if irqfd used with no in-kernel XICS emulation
Bugzilla: 1375778
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

From: Paul Mackerras <paulus@ozlabs.org>

It turns out that if userspace creates a pseries-type VM without
in-kernel XICS (interrupt controller) emulation, and then connects
an eventfd to the VM as an irqfd, and the eventfd gets signalled,
that the code will try to deliver an interrupt via the non-existent
XICS object and crash the host kernel with a NULL pointer dereference.

To fix this, we check for the presence of the XICS object before
trying to deliver the interrupt, and return with an error if not.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
(cherry picked from commit e48ba1cbce12eb4546771d45c09dd94c3404efe8)

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1375778

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>