[kernel] perf/core: Free AUX pages in unmap path
Message-id: <1474556776-20705-4-git-send-email-jolsa@redhat.com>
Patchwork-id: 158071
O-Subject: [PATCH RHEL7.3 BZ1373817 003/252] perf/core: Free AUX pages in unmap path
Bugzilla: 1373817
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1373817
https://bugzilla.redhat.com/show_bug.cgi?id=1373817

upstream
========
commit 95ff4ca26c492fc1ed7751f5dd7ab7674b54f4e0
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Wed Dec 2 18:41:11 2015 +0200

description
===========
Now that we can ensure that when ring buffer's AUX area is on the way
to getting unmapped new transactions won't start, we only need to stop
all events that can potentially be writing aux data to our ring buffer.

Having done that, we can safely free the AUX pages and corresponding
PMU data, as this time it is guaranteed to be the last aux reference
holder.

This partially reverts:

  57ffc5ca679 ("perf: Fix AUX buffer refcounting")

... which was made to defer deallocation that was otherwise possible
from an NMI context. Now it is no longer the case; the last call to
rb_free_aux() that drops the last AUX reference has to happen in
perf_mmap_close() on that AUX area.

Signed-off-by: Rafael Aquini <aquini@redhat.com>