[tools] perf tools: Add template for generating rbtree resort class
Message-id: <1474556776-20705-169-git-send-email-jolsa@redhat.com>
Patchwork-id: 158230
O-Subject: [PATCH RHEL7.3 BZ1373817 168/252] perf tools: Add template for generating rbtree resort class
Bugzilla: 1373817
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1373817
https://bugzilla.redhat.com/show_bug.cgi?id=1373817

upstream
========
commit f58c253564815db541e93b5411c2b47dbcb2f0ea
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu May 5 11:44:28 2016 -0300

description
===========
Sometimes we want to sort an existing rbtree by a different key,
introduce a template for that, that needs only to be provided the
rbtree root and the number of entries in it.

To do that a new rbtree will be created with extra space for each entry,
where possibly pre-calculated keys will be stored to be used in the
resort process and also later, when using the newly sorted rbtree.

Please check the following two changesets to see it in use for resorting
stats for threads and its syscalls in 'perf trace --summary'.

Signed-off-by: Rafael Aquini <aquini@redhat.com>