[tools] perf powerpc: Fix kprobe and kretprobe handling with kallsyms on ppc64le
Message-id: <1474556776-20705-181-git-send-email-jolsa@redhat.com>
Patchwork-id: 158246
O-Subject: [PATCH RHEL7.3 BZ1373817 180/252] perf powerpc: Fix kprobe and kretprobe handling with kallsyms on ppc64le
Bugzilla: 1373817 1376534
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1373817
https://bugzilla.redhat.com/show_bug.cgi?id=1373817

upstream
========
commit 239aeba764092b29dd7cab177cd47f472390622e
Author: Naveen N. Rao <naveen.n.rao@linux.vnet.ibm.com>
Date: Tue Apr 12 14:40:49 2016 +0530

description
===========
So far, we used to treat probe point offsets as being offset from the
LEP. However, userspace applications (objdump/readelf) always show
disassembly and offsets from the function GEP. This is confusing to the
user as we will end up probing at an address different from what the
user expects when looking at the function disassembly with
readelf/objdump. Fix this by changing how we modify probe address with
perf.

If only the function name is provided, we assume the user needs the LEP.
Otherwise, if an offset is specified, we assume that the user knows the
exact address to probe based on function disassembly, and so we just
place the probe from the GEP offset.

Finally, kretprobe was also broken with kallsyms as we were trying to
specify an offset. This patch also fixes that issue.

Signed-off-by: Rafael Aquini <aquini@redhat.com>