[tools] perf trace: Do not print raw args list for syscalls with no args
Message-id: <1474556776-20705-185-git-send-email-jolsa@redhat.com>
Patchwork-id: 158247
O-Subject: [PATCH RHEL7.3 BZ1373817 184/252] perf trace: Do not print raw args list for syscalls with no args
Bugzilla: 1373817
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1373817
https://bugzilla.redhat.com/show_bug.cgi?id=1373817

upstream
========
commit 4c4d6e519091a3551a0ad2e2e8423d77f3000400
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu May 5 23:38:05 2016 -0300

description
===========
The test to check if the arg format had been read from the
syscall:sys_enter_name/format file was looking at the list of non-commom
fields, and if that is empty, it would think it had failed to read it,
because it doesn't exist, for instance, for the clone() syscall.

So instead before dumping the raw syscall args list check
IS_ERR(sc->tp_format), if that is true, then an attempt was made to read
the format file and failed, in which case dump the raw arg list values.

Signed-off-by: Rafael Aquini <aquini@redhat.com>