[tools] perf symbols: Store vdso buildid unconditionally
Message-id: <1474556776-20705-228-git-send-email-jolsa@redhat.com>
Patchwork-id: 158291
O-Subject: [PATCH RHEL7.3 BZ1373817 227/252] perf symbols: Store vdso buildid unconditionally
Bugzilla: 1373817
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1373817
https://bugzilla.redhat.com/show_bug.cgi?id=1373817

upstream
========
commit 6ae98ba611ed1c11ddc5645475bc03b46a3c04e7
Author: He Kuang <hekuang@huawei.com>
Date: Thu May 12 08:43:11 2016 +0000

description
===========
When unwinding callchains on a different machine, vdso info should be
available so the unwind process won't be interrupted if address falls
into vdso region. But in most cases, the addresses of sample events are
not in vdso range, the buildid of a zero hit vdso won't be stored into
perf.data.

This patch stores vdso buildid regardless of whether the vdso is hit or
not.

Signed-off-by: Rafael Aquini <aquini@redhat.com>