[tools] perf machine: Do not bail out if not managing to read ref reloc symbol
Message-id: <1474556776-20705-233-git-send-email-jolsa@redhat.com>
Patchwork-id: 158295
O-Subject: [PATCH RHEL7.3 BZ1373817 232/252] perf machine: Do not bail out if not managing to read ref reloc symbol
Bugzilla: 1373817
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1373817
https://bugzilla.redhat.com/show_bug.cgi?id=1373817

upstream
========
commit 45e90056904b12d8dd74e0d2ea6dfd5e4394104d
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue May 17 11:52:26 2016 -0300

description
===========
This means the user can't access /proc/kallsyms, for instance, because
/proc/sys/kernel/kptr_restrict is set to 1.

Instead leave the ref_reloc_sym as NULL and code using it will cope.

This allows 'perf trace' to work on such systems for !root, the only
issue would be when trying to resolve kernel symbols, which happens,
for instance, in some libtracevent plugins.  A warning for that case
will be provided in the next patch in this series.

Noticed in Ubuntu 16.04, that comes with kptr_restrict=1.

Signed-off-by: Rafael Aquini <aquini@redhat.com>