[tools] perf trace: Use the fd->name beautifier as default for "fd" args
Message-id: <1474556776-20705-244-git-send-email-jolsa@redhat.com>
Patchwork-id: 158306
O-Subject: [PATCH RHEL7.3 BZ1373817 243/252] perf trace: Use the fd->name beautifier as default for "fd" args
Bugzilla: 1373817
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1373817
https://bugzilla.redhat.com/show_bug.cgi?id=1373817

upstream
========
commit b6565c908ad7eb28dfdda9578ec5a074e080cedc
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon May 23 12:59:53 2016 -0300

description
===========
Noticed when the 'setsockopt' 'fd' arg wasn't being formatted via
the SCA_FD beautifier, so just remove the setting of "fd" args to
SCA_FD and do it when reading the syscall info, like we do for
args of type "pid_t", i.e. "fd" as the name should be enough as
the decision to use the SFA_FD beautifier. For odd cases we can
just do it explicitely.

Signed-off-by: Rafael Aquini <aquini@redhat.com>